- [ ] Changelog updated / no changelog update needed

+++ agb-image-converter/src/sprite/regular.rs
@@ -145,7 +145,7 @@ impl ToTokens for Output {
             let palette_idx = sprite.palette as usize;
 
             quote! {
-                unsafe { Sprite::new(&PALETTES[#palette_idx], align_bytes!(u16, #data), Size::from_width_height(#x, #y)) }
+                unsafe { Sprite::new(&PALETTES[#palette_idx], align_bytes!(u32, #data), Size::from_width_height(#x, #y)) }
             }
         });
